Move keysinterface.rs to a directory-level module called sign.
