+++ fuzz/src/invoice_request_deser.rs
@@ -11,7 +11,7 @@ use bitcoin::secp256k1::{KeyPair, Parity, PublicKey, Secp256k1, SecretKey, self}
 use crate::utils::test_logger;
 use core::convert::{Infallible, TryFrom};
 use lightning::blinded_path::BlindedPath;
-use lightning::chain::keysinterface::EntropySource;
+use lightning::sign::EntropySource;
 use lightning::ln::PaymentHash;
 use lightning::ln::features::BlindedHopFeatures;
 use lightning::offers::invoice::{BlindedPayInfo, UnsignedInvoice};
