+++ lightning/src/events/mod.rs
@@ -20,7 +20,7 @@ pub mod bump_transaction;
 #[cfg(anchors)]
 pub use bump_transaction::BumpTransactionEvent;
 
-use crate::chain::keysinterface::SpendableOutputDescriptor;
+use crate::sign::SpendableOutputDescriptor;
 use crate::ln::channelmanager::{InterceptId, PaymentId, RecipientOnionFields};
 use crate::ln::channel::FUNDING_CONF_DEADLINE_BLOCKS;
 use crate::ln::features::ChannelTypeFeatures;
@@ -129,7 +129,7 @@ pub enum ClosureReason {
 		/// Be careful about printing the peer_msg, a well-crafted message could exploit
 		/// a security vulnerability in the terminal emulator or the logging subsystem.
 		/// To be safe, use `Display` on `UntrustedString`
-		/// 
+		///
 		/// [`UntrustedString`]: crate::util::string::UntrustedString
 		peer_msg: UntrustedString,
 	},
@@ -377,7 +377,7 @@ pub enum Event {
 		/// This field will always be filled in when the event was generated by LDK versions
 		/// 0.0.113 and above.
 		///
-		/// [phantom nodes]: crate::chain::keysinterface::PhantomKeysManager
+		/// [phantom nodes]: crate::sign::PhantomKeysManager
 		receiver_node_id: Option<PublicKey>,
 		/// The hash for which the preimage should be handed to the ChannelManager. Note that LDK will
 		/// not stop you from registering duplicate payment hashes for inbound payments.
@@ -425,7 +425,7 @@ pub enum Event {
 		/// This field will always be filled in when the event was generated by LDK versions
 		/// 0.0.113 and above.
 		///
-		/// [phantom nodes]: crate::chain::keysinterface::PhantomKeysManager
+		/// [phantom nodes]: crate::sign::PhantomKeysManager
 		receiver_node_id: Option<PublicKey>,
 		/// The payment hash of the claimed payment. Note that LDK will not stop you from
 		/// registering duplicate payment hashes for inbound payments.
