+++ lightning/src/util/errors.rs
@@ -61,7 +61,7 @@ pub enum APIError {
 	/// Using a SegWit v0 script should resolve this issue. If you cannot, you won't be able to open
 	/// a channel or cooperatively close one with this peer (and will have to force-close instead).
 	///
-	/// [`SignerProvider::get_shutdown_scriptpubkey`]: crate::chain::keysinterface::SignerProvider::get_shutdown_scriptpubkey
+	/// [`SignerProvider::get_shutdown_scriptpubkey`]: crate::sign::SignerProvider::get_shutdown_scriptpubkey
 	/// [`InitFeatures`]: crate::ln::features::InitFeatures
 	IncompatibleShutdownScript {
 		/// The incompatible shutdown script.
